requirements.txt

Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
gunicorn==20.1.0
Markdown==3.3.7
mysql-connector==2.2.9
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
virtualenv==20.14.1
whitenoise==6.1.0
